mpegmap/HTMLSourcingInbandTracks-3.3-3.html

<!doctype html>
<html>
  <head>
    <title>TC 5.1.4-5 Program Insertion Cue TextTrack</title>
    <link rel="author" title="CableLabs" href="html5@cablelabs.com">
    <meta name="flags" content="[requirement flags]">
    <meta name="assert" content="MPEG-2 SPTS, Program Insertion Cue - UA creates TextTrack, attributes set.">
    <meta name="timeout" content="long">
    <script src="testmedia.js"></script>
    <!-- Source HTTP server local copy of World Wide Web Consortium (W3C) Test Harness. --> 
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <!-- Script Version: BETA -->
    <!-- About the Media Resource

         To verify DOM elements the details of the test media must be known prior to test execution.
         This test script is bound to the test media file sourced in the script.
         Sourcing a differnt test media file, or modifying that file, may invalidate the test.

         The media is an MPEG-2, Single Program Transport Stream
         containing two EISS streams, one program insertion cue stream.
    -->
  </head>
  <body>
    <p>
      <a href="http://www.w3.org/TR/html5/references.html#refsINBANDTRACKS">Spec
      Reference [INBANDTRACKS]</a>
    </p>
    <video id="v" width="400" height="300" autoplay controls></video>
    <div id="log"></div>
    <script>
      var test = async_test("Program Insertion Cue TextTrack attributes.");
      var vid = document.getElementById("v");
      vid.src = mediaServerURL() + "UserPrivateStreams.ts"; 

      var pid = '45'; // Program Insertion Cue Message PID

      vid.textTracks.addEventListener("addtrack", function(event) {
        var track = event.track;
        if (track.id === pid) {
          test.step(function() {
            assert_equals(track.kind, "metadata", "TextTrack.kind");
            assert_equals(track.language, "", "TextTrack.language");
            assert_equals(track.mode, "disabled", "TextTrack.mode");
          });
          test.done();
        }
      });

      vid.addEventListener("playing", function() {
        vid.pause();
        if (!vid.textTracks.getTrackById(pid)) {
          test.step(function() {
            assert_unreached("Did not find  Program Insertion Cue TextTrack.id (PID) " + pid + " in DOM.");
          }, "Did not find Program Insertion Cue TextTrack.id (PID) " + pid + " in DOM.");
          test.done();
        }
      });
    </script>
  </body>
</html>
